+++ src/main/java/com/bobrust/util/Sign.java
@@ -1,4 +1,4 @@
-package com.bobrust.gui;
+package com.bobrust.util;
 
 import java.awt.Color;
